名刺管理アプリ_アップロード用/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
sqlalchemy==2.0.38
pydantic==2.10.6
python-multipart==0.0.20
google-generativeai==0.8.4
python-dotenv==1.0.1
Pillow==11.1.0
watchdog==6.0.0
PyMuPDF==1.25.3
pyngrok==7.1.6
supabase==2.3.4
psycopg2-binary==2.9.9
